moved the json converter into a test context
